crc32: add real 8 bit
crc32.c provides a choice of one of several algorithms for computing
the LSB and LSB versions of the CRC32 checksum based on the parameters
CRC_LE_BITS and CRC_BE_BITS. In the original version the
values 1, 2, 4 and 8 respectively selected versions of the
alrogithm that computed the crc 1, 2, 4 and 32 bits as a
time. This patch series adds a new version that computes
the CRC 64 bits at a time. To make things easier to
understand the parameter has been reinterpreted to actually
stand for the number of bits processed in each step of the
algorithm so that the old value 8 has been replaced with
the value 32. This also allows us to add in a widely used
crc algorithm that computes the crc 8 bits at a time called
the Sarwate algorithm.

Signed-off-by: Bob Pearson <rpearson@systemfabricworks.com>